requirements.txt
# Core dependencies
simpy>=4.0.0
networkx>=3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
torch>=2.0.0
gym>=0.26.0

# Topology generation
scikit-learn>=1.3.0
python-louvain>=0.16

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
psutil>=5.9.0
tqdm>=4.65.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Optional (for BRITE integration)
# Note: BRITE requires Java and is not a Python package
# Recommendation: use openjdk-17-jdk